linguistic-systems/movie-chat-bot/requirements.txt
pip~=24.0
emoji~=2.14.1
attrs~=24.2.0
protobuf~=6.31.0
docutils~=0.20.1
Sphinx~=7.2.5
Jinja2~=3.1.4
setuptools~=65.5.0
typing_extensions~=4.12.2
platformdirs~=4.2.2
python-telegram-bot~=22.5
openai~=1.3.0
pandas~=2.1.0
numpy~=1.24.0
scikit-learn~=1.3.0
python-dotenv~=1.0.0
gigachat~=0.1.42.post2
dotenv~=0.9.9
